style: Move Member entity and dto to member.entity package and member.dto package
