+++ src/main/java/gift/member/MemberRepository.java
@@ -1,5 +1,6 @@
 package gift.member;
 
+import gift.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
