+++ src/main/java/gift/member/entity/Member.java
@@ -1,4 +1,4 @@
-package gift.member;
+package gift.member.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
