+++ src/test/java/gift/member/MemberRepositoryTest.java
@@ -3,6 +3,7 @@ package gift.member;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import gift.member.entity.Member;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
